requirements.txt
fastapi
uvicorn[standard]
langgraph
langchain-google-genai
google-genai
python-dotenv
pydantic
python-multipart
pillow
